refactor : remove compiler warnings

+++ JungleTechLab_Week3/FEditorViewportClient.cpp
@@ -59,7 +59,7 @@ void FEditorViewportClient::RayCast(D3D11_VIEWPORT ViewportInfo, UWorld* World, 
 	//	DeprojectScreenToWorldForOrtho(WindowApplication.Input.CursorX - ViewportInfo.TopLeftX, WindowApplication.Input.CursorY - ViewportInfo.TopLeftY,
 	//		ViewportInfo.Width, ViewportInfo.Height, 0.1f, 100.f, NearPoint, FarPoint);
 	//}
-	DeprojectScreenToWorldForUnified(WindowApplication.Input.CursorX - ViewportInfo.TopLeftX, WindowApplication.Input.CursorY - ViewportInfo.TopLeftY,
+	DeprojectScreenToWorldForUnified(WindowApplication.Input.CursorX - static_cast<int32>(ViewportInfo.TopLeftX), static_cast<int32>(WindowApplication.Input.CursorY) - ViewportInfo.TopLeftY,
 		ViewportInfo.Width, ViewportInfo.Height, 0.1f, 100.f, mCamera.mOrthoDistance, perspectiveRatio, NearPoint, FarPoint);
 
 	mRayNear = NearPoint;
